+++ pybillomat/__init__.py
@@ -1,0 +1,8 @@
+#!/usr/bin/env python
+# coding: utf-8
+"""
+Python-Billomat - Billomat API Client Library
+"""
+
+from .connection import Connection
+from .clients import Client, Clients