+++ pybillomat/clients.py
@@ -1,0 +1,215 @@
+#!/usr/bin/env python
+# coding: utf-8
+"""
+Clients
+"""
+
+import datetime
+import xml.etree.ElementTree as etree
+from bunch import Bunch
+
+
+class Client(Bunch):
+
+    content_language = None
+    id = None  # Integer
+    created = None  # Datetime
+    archived = None  # Boolean
+    client_number = None
+    number = None  # Integer
+    number_pre = None
+    name = None
+    salutation = None
+    first_name = None
+    last_name = None
+    street = None
+    zip = None
+    city = None
+    state = None
+    country_code = None
+    phone = None
+    fax = None
+    mobile = None
+    email = None
+    www = None
+    tax_number = None
+    vat_number = None
+    bank_account_owner = None
+    bank_number = None
+    bank_name = None
+    bank_account_number = None
+    bank_swift = None
+    bank_iban = None
+    sepa_mandate = None
+    sepa_mandate_date = None
+    tax_rule = None
+    net_gross = None
+    default_payment_types = None
+    discount_rate_type = None
+    discount_rate = None
+    discount_days_type = None
+    discount_days = None
+    due_days_type = None
+    due_days = None
+    reminder_due_days_type = None
+    reminder_due_days = None
+    offer_validity_days_type = None
+    offer_validity_days = None
+    currency_code = None
+    price_group = None
+    note = None
+    revenue_gross = None  # Float
+    revenue_net = None  # Float"
+
+
+    def __init__(self, conn):
+        self.conn = conn
+
+
+    @classmethod
+    def from_etree(cls, conn, etree_element):
+        """
+        Returns Client-object from ElementTree-Element
+        """
+
+        client = cls(conn = conn)
+
+        for item in etree_element:
+
+            # Get data
+            isinstance(item, etree.Element)
+            tag = item.tag
+            type = item.attrib.get("type")
+            text = item.text
+
+            if not text is None:
+                if type == "integer":
+                    setattr(client, tag, int(text))
+                elif type == "datetime":
+                    # <created type="datetime">2011-10-04T17:40:00+02:00</created>
+                    dt = datetime.datetime.strptime(text[:19], "%Y-%m-%dT%H:%M:%S")
+                    setattr(client, tag, dt)
+                else:
+                    setattr(client, tag, text)
+
+            # <plan>L</plan>
+            # <quotas>
+                # <quota>
+                    # <entity>documents</entity>
+                    # <available>300</available>
+                    # <used>22</used>
+                # </quota>
+                # <quota>
+                    # <entity>clients</entity>
+                    # <available>1500</available>
+                    # <used>328</used>
+                # </quota>
+                # <quota>
+                    # <entity>articles</entity>
+                    # <available>5000</available>
+                    # <used>16</used>
+                # </quota>
+                # <quota>
+                    # <entity>storage</entity>
+                    # <available>-1</available>
+                    # <used>203782576</used>
+                # </quota>
+            # </quotas>
+
+        # Finished
+        return client
+
+
+    @classmethod
+    def from_xml(cls, conn, xml_string):
+        """
+        Returns new Client-object from XML-string
+        """
+
+        # Parse XML
+        root = etree.fromstring(xml_string)
+
+        # Finished
+        return cls.from_etree(conn, root)
+
+
+    @classmethod
+    def get(cls, conn, id = None):
+        """
+        Returns Client-object
+        """
+
+        # Path
+        path = "/api/clients/{id}".format(id = id)
+
+        # Request
+        request = conn.request(method = "GET", url = path)
+
+        # Create new client-object from XML
+        client = cls.from_xml(conn, request.data)
+        client.content_language = request.headers.get("content-language", None)
+
+        # Finished
+        return client
+
+
+class Clients(object):
+
+
+    def __init__(
+        self,
+        conn
+    ):
+        self.conn = conn
+
+
+    @classmethod
+    def from_search(
+        cls,
+        conn,
+        name = None,
+        client_number = None,
+        email = None,
+        first_name = None,
+        last_name = None,
+        country_code = None,
+        note = None,
+        invoice_id = None,
+        tags = None
+    ):
+        """
+        Returns Client-object
+
+        :param name: Company name
+        :param client_number: Client number
+        :param email: E-mail address
+        :param first_name: First name of the contact person
+        :param last_name: Last name of the contact person
+        :param country_code: Country code as ISO 3166 Alpha-2
+        :param note: Note
+        :param invoice_id: ID of an invoice of this client,
+            multiple values seperated with comma
+        :param tags: Comma seperated list of tags
+        """
+
+        # Path
+        path = "/api/clients"
+
+        # Suchparameter
+        if name:
+            path += "?name={name}".format(name = name)
+        # ...
+
+
+
+
+        # # Request
+        # request = conn.request(method = "GET", url = path)
+        #
+        # # Create new client-object from XML
+        # client = cls.from_xml(conn, request.data)
+        # client.content_language = request.headers.get("content-language", None)
+        #
+        # # Finished
+        # return client
+