+++ history.rst
@@ -6,6 +6,19 @@ by Gerold Penz 2014
 
 
 =============
+Version 0.1.1
+=============
+
+2014-10-26
+
+- Tests with *urllib3*
+
+- Connection-module added. It uses *urllib3* to connect to Billomat.
+
+- Clients-module added.
+
+
+=============
 Version 0.0.2
 =============